Integrated Completed Project From Josh's email
This change will (hopefully) be the the final integration to complete
our project. I have added everything from Josh, everything that I have
been working on with the colors and the menus, and everything that Dori
has been working on with the account class.

This is bound to cause problems so I will make sure that all the testing
that happens happens to this one file.