Prospective Customers
Here are the files for the Prospective customers, including a queue
template that is used to list them all, and necessary headers. The stuff
in ProspectTemplateHeader.h can pretty much just get shoved into
whatever main header file we have. It's not working perfectly right now
and I'm not sure why, but it seems to be something in the InfoRequest
function in ProspectFunctions.cpp. Not sure if I'm using the try catch
wrong or what, but if any of you can see something I'm missing, let me
know.